requirements.txt
autopep8==2.0.0
mypy==0.991
mypy-extensions==0.4.3
pycodestyle==2.10.0
python-dotenv==0.21.0
tomli==2.0.1
typing_extensions==4.4.0
